requirements.txt
django-bootstrap5==22.2
requests